requirements.txt
streamlit~=1.35.0
biopython~=1.83